backend/requirements.txt
asgiref==3.2.5
astroid==2.3.3
certifi==2020.4.5.1
Django==3.0.4
django-debug-toolbar==2.2
django-environ==0.4.5
django-model-utils==4.0.0
django-safedelete==0.5.4
djangorestframework==3.11.0
dry-rest-permissions==0.1.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
olefile==0.46
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
python-http-client==3.2.6
pytz==2019.3
PyYAML==5.3.1
sendgrid==6.2.0
six==1.14.0
sqlparse==0.3.1
wrapt==1.12.1
